Added IsPascalCase method and updated version number to 1.2.2
In StringExtension.cs file, under IceCoffee.Common.Extensions namespace, added a new method IsPascalCase. This method is used to determine whether a word is in Pascal Case format, that is, the first letter is capitalized and there is at least one lowercase letter in the subsequent letters.
The method for converting the input string to Pascal Case has been modified to add a check whether the word is already in Pascal Case format. If so, skip the conversion.
In IceCoffee.Common.csproj file, updated the version number from 1.2.1 to 1.2.2.