Added datetime extension methods and updated version numbers
In the `StringExtension.cs` file, added `ToDateTime` and `ToDateTimeNullable` extension methods for converting strings to `DateTime` and nullable `DateTime`, and returning default values ​​when the format is wrong. Also added corresponding XML comments.

In the `IceCoffee.Common.csproj` file, updated the version number from `1.3.0` to `1.3.1`, and the rest of the project properties remain unchanged.